Score_Algorithm_rev04

+++ src/com/xmiles/android/Gmaps_backup_20161001.java
@@ -18,6 +18,7 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.android.maps.GeoPoint;
+import com.xmiles.android.backup.FbPlaces_Download;
 import com.xmiles.android.sqlite.contentprovider.SqliteProvider;
 import com.xmiles.android.sqlite.helper.DatabaseHelper;
 import com.xmiles.android.support.GPSTracker;
@@ -25,7 +26,6 @@ import com.xmiles.android.support.GetDistance;
 import com.xmiles.android.support.Score_Algorithm;
 import com.xmiles.android.support.Support;
 import com.xmiles.android.webservice.UserFunctions;
-import com.xmiles.android.scheduler.FbPlaces_Download;
 import com.xmiles.android.scheduler.NewsFeed_Inbox_Upload;
 
 import com.xmiles.android.scheduler.Getting_GpsBusData;
@@ -248,8 +248,8 @@ public class Gmaps_backup_20161001 extends FragmentActivity {
       					                (int) (gps.getLatitude()  * 1E6),
       					                (int) (gps.getLongitude() * 1E6));
 
-      						    float Lat    = (float) (curGeoPoint.getLatitudeE6() / 1E6);
-      						    float Long   = (float) (curGeoPoint.getLongitudeE6() / 1E6);
+      							double Lat    = (double) (curGeoPoint.getLatitudeE6() / 1E6);
+      							double Long   = (double) (curGeoPoint.getLongitudeE6() / 1E6);
 
       							
       							Double get_distance =  distance.calculo(Double.parseDouble(dataBusArray[index_LATITUDE]),
